FactorAdditionOp is now able to add a vector of drwnGeneralFactors (lazily).
Changes to drwnGeneralFactor:
	removed addVariable (inherit from drwnFactor)
	added copy constructor
	added clone
	added getValueOf that has parameter drwnFullAssignment
	added setStorageType, which is required as a FactorAdditionOp determines
the storage type of its target in constructor but delays this assignment to the
target until execute is called. Consider how we may want to add two sparse
factors (over the same set of variables) but won't now the density of the sum.

Modified drwnTableFactorNAryOp's initialize to cater for drwnGeneralFactors.